add support for sending mortal txs (#43)
This PR takes the mortal flag value passed in CLI and passes it over up
to subxt transaction building, where we can configure its tx params to
contain the mortality.

Closes #14 

### Review notes

TODO:
- ~test this work by writing a scenario in polkadot-sdk and see how
mortal txs work~ (works after using subxt correctly)
- ~look for unit tests that might be implemented in the repo~ (not that
straight forward, useful thing to test is mortal/immortal txs creation,
but we have integration tests in `polkadot-sdk` that covers it)
- ~maybe enable support for mixing mortal with immortal txs?~ (not
really needed for now)

---------

+++ src/cli.rs
@@ -30,9 +30,10 @@ pub enum CliCommand {
 		/// Spawn block monitor for checking if transactions are included in finalized blocks.
 		#[clap(long)]
 		block_monitor: bool,
-		/// Use mortal transactions.
+		/// Use mortal transactions. This represents the number of blocks the mortal tx is valid
+		/// for, starting with the current finalized block at the time of tx creation.
 		#[clap(long)]
-		mortal: Option<u32>,
+		mortal: Option<u64>,
 		/// Send transactions threshold, sends the batch when number of pedning extrinsics drops
 		/// below this number.
 		#[clap(long, default_value_t = 10000)]
